+++ lang/src/accounts/mod.rs
@@ -9,9 +9,6 @@ pub mod boxed;
 pub mod cpi_account;
 #[doc(hidden)]
 #[allow(deprecated)]
-pub mod cpi_state;
-#[doc(hidden)]
-#[allow(deprecated)]
 pub mod loader;
 pub mod option;
 pub mod program;
@@ -19,9 +16,6 @@ pub mod program;
 #[allow(deprecated)]
 pub mod program_account;
 pub mod signer;
-#[doc(hidden)]
-#[allow(deprecated)]
-pub mod state;
 pub mod system_account;
 pub mod sysvar;
 pub mod unchecked_account;