+++ ts/packages/anchor/src/coder/index.ts
@@ -19,11 +19,6 @@ export interface Coder<A extends string = string, T extends string = string> {
   readonly accounts: AccountsCoder<A>;
 
   /**
-   * Coder for state structs.
-   */
-  readonly state: StateCoder;
-
-  /**
    * Coder for events.
    */
   readonly events: EventCoder;
@@ -49,7 +44,6 @@ export interface AccountsCoder<A extends string = string> {
 
 export interface InstructionCoder {
   encode(ixName: string, ix: any): Buffer;
-  encodeState(ixName: string, ix: any): Buffer;
 }
 
 export interface EventCoder {