+++ CHANGELOG.md
@@ -11,11 +11,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 ## [Unreleased]
 
 ### Features
+
 - cli: Add `env` option to verifiable builds ([#2325](https://github.com/coral-xyz/anchor/pull/2325)).
 
 ### Fixes
 
 ### Breaking
+
+- lang: Remove `state` and `interface` attributes ([#2285](https://github.com/coral-xyz/anchor/pull/2285)).
 
 ## [0.26.0] - 2022-12-15